♻️ Move zIndex constants and columnHandleId utility to ERDRenderer, and refactor imports in related components
