+++ .changeset/thirty-pants-invent.md
@@ -1,0 +1,5 @@
+---
+"@liam-hq/erd-core": patch
+---
+
+♻️ Move zIndex constants and columnHandleId utility to ERDRenderer, and refactor imports in related components
