+++ frontend/packages/erd-core/src/components/ERDRenderer/components/ERDContent/ERDContent.tsx
@@ -1,7 +1,6 @@
 import { selectTableLogEvent } from '@/features/gtm/utils'
 import { repositionTableLogEvent } from '@/features/gtm/utils/repositionTableLogEvent'
 import { MAX_ZOOM, MIN_ZOOM } from '@/features/reactflow/constants'
-import { highlightNodesAndEdges } from '@/features/reactflow/utils'
 import { useIsTouchDevice } from '@/hooks'
 import { useVersion } from '@/providers'
 import { useUserEditingActiveStore } from '@/stores'
@@ -19,6 +18,7 @@ import {
 import { type FC, useCallback } from 'react'
 import { useTableSelection } from '../../hooks'
 import type { DisplayArea } from '../../types'
+import { highlightNodesAndEdges } from '../../utils'
 import styles from './ERDContent.module.css'
 import { ERDContentProvider, useERDContentContext } from './ERDContentContext'
 import { NonRelatedTableGroupNode } from './NonRelatedTableGroupNode'
