+++ frontend/packages/erd-core/src/components/ERDRenderer/components/LeftPane/TableNameMenuButton/TableNameMenuButton.tsx
@@ -1,6 +1,6 @@
 import { useTableSelection } from '@/components/ERDRenderer/hooks'
+import type { TableNodeType } from '@/components/ERDRenderer/types'
 import { selectTableLogEvent } from '@/features/gtm/utils'
-import type { TableNodeType } from '@/features/reactflow/types'
 import { useVersion } from '@/providers'
 import { SidebarMenuButton, SidebarMenuItem, Table2 } from '@liam-hq/ui'
 import clsx from 'clsx'
