+++ frontend/packages/erd-core/src/components/ERDRenderer/constants.ts
@@ -1,0 +1,5 @@
+export const zIndex = {
+  nodeDefault: 2,
+  edgeHighlighted: 1,
+  edgeDefault: 0,
+}
