+++ frontend/packages/erd-core/src/components/ERDRenderer/types.ts
@@ -1,1 +1,16 @@
+import type { Cardinality, Table } from '@liam-hq/db-structure'
+import type { Node } from '@xyflow/react'
+
+export type TableNodeData = {
+  table: Table
+  isActiveHighlighted: boolean
+  isHighlighted: boolean
+  sourceColumnName: string | undefined
+  targetColumnCardinalities?:
+    | Record<string, Cardinality | undefined>
+    | undefined
+}
+
+export type TableNodeType = Node<TableNodeData, 'table'>
+
 export type DisplayArea = 'main' | 'relatedTables'
