+++ frontend/packages/erd-core/src/components/ERDRenderer/utils/columnHandleId.ts
