+++ frontend/packages/erd-core/src/components/ERDRenderer/utils/highlightNodesAndEdges.test.ts
