+++ frontend/packages/erd-core/src/components/ERDRenderer/utils/highlightNodesAndEdges.ts
@@ -1,4 +1,4 @@
-import { zIndex } from '@/features/reactflow/constants'
+import { zIndex } from '@/components/ERDRenderer/constants'
 import type { Edge, Node } from '@xyflow/react'
 import type { TableNodeType } from '../types'
 import { isTableNode } from './isTableNode'
