+++ frontend/packages/erd-core/src/components/ERDRenderer/utils/index.ts
@@ -1,1 +1,5 @@
 export * from './computeAutoLayout'
+export * from './columnHandleId'
+export * from './convertDBStructureToNodes'
+export * from './highlightNodesAndEdges'
+export * from './isTableNode'
