+++ frontend/packages/erd-core/src/components/ERDRenderer/utils/isTableNode.ts
