+++ frontend/packages/erd-core/src/features/reactflow/constants.ts
@@ -1,8 +1,2 @@
 export const MIN_ZOOM = 0.1
 export const MAX_ZOOM = 2
-
-export const zIndex = {
-  nodeDefault: 2,
-  edgeHighlighted: 1,
-  edgeDefault: 0,
-}
