--- frontend/packages/erd-core/src/features/reactflow/types.ts
@@ -1,14 +1,0 @@
-import type { Cardinality, Table } from '@liam-hq/db-structure'
-import type { Node } from '@xyflow/react'
-
-export type TableNodeData = {
-  table: Table
-  isActiveHighlighted: boolean
-  isHighlighted: boolean
-  sourceColumnName: string | undefined
-  targetColumnCardinalities?:
-    | Record<string, Cardinality | undefined>
-    | undefined
-}
-
-export type TableNodeType = Node<TableNodeData, 'table'>
